Unwatch dropped channel when init channel manager (#11986)
issue: #11558
