+++ internal/datacoord/server.go
@@ -28,7 +28,6 @@ import (
 	datanodeclient "github.com/milvus-io/milvus/internal/distributed/datanode/client"
 	rootcoordclient "github.com/milvus-io/milvus/internal/distributed/rootcoord/client"
 	"github.com/milvus-io/milvus/internal/logutil"
-	"github.com/milvus-io/milvus/internal/rootcoord"
 	"github.com/milvus-io/milvus/internal/util/metricsinfo"
 	"github.com/milvus-io/milvus/internal/util/mqclient"
 	"github.com/milvus-io/milvus/internal/util/tsoutil"
@@ -46,7 +45,6 @@ import (
 
 	"github.com/milvus-io/milvus/internal/proto/commonpb"
 	"github.com/milvus-io/milvus/internal/proto/datapb"
-	"github.com/milvus-io/milvus/internal/proto/internalpb"
 	"github.com/milvus-io/milvus/internal/proto/milvuspb"
 )
 
@@ -89,9 +87,6 @@ type rootCoordCreatorFunc func(ctx context.Context, metaRootPath string, etcdEnd
 // makes sure Server implements `DataCoord`
 var _ types.DataCoord = (*Server)(nil)
 
-// makes sure Server implements `positionProvider`
-var _ positionProvider = (*Server)(nil)
-
 // Server implements `types.Datacoord`
 // handles Data Cooridinator related jobs
 type Server struct {
@@ -112,6 +107,7 @@ type Server struct {
 	rootCoordClient  types.RootCoord
 	garbageCollector *garbageCollector
 	gcOpt            GcOption
+	handler          Handler
 
 	compactionTrigger trigger
 	compactionHandler compactionPlanContext
@@ -248,6 +244,8 @@ func (s *Server) Start() error {
 		return err
 	}
 
+	s.handler = newServerHandler(s)
+
 	if err = s.initCluster(); err != nil {
 		return err
 	}
@@ -282,7 +280,7 @@ func (s *Server) initCluster() error {
 	}
 
 	var err error
-	s.channelManager, err = NewChannelManager(s.kvClient, s)
+	s.channelManager, err = NewChannelManager(s.kvClient, s.handler)
 	if err != nil {
 		return err
 	}
@@ -782,97 +780,4 @@ func (s *Server) loadCollectionFromRootCoord(ctx context.Context, collectionID i
 	}
 	s.meta.AddCollection(collInfo)
 	return nil
-}
-
-// GetVChanPositions get vchannel latest postitions with provided dml channel names
-func (s *Server) GetVChanPositions(channel string, collectionID UniqueID, partitionID UniqueID) *datapb.VchannelInfo {
-	segments := s.meta.GetSegmentsByChannel(channel)
-	log.Debug("GetSegmentsByChannel",
-		zap.Any("collectionID", collectionID),
-		zap.Any("channel", channel),
-		zap.Any("numOfSegments", len(segments)),
-	)
-	var flushed []*datapb.SegmentInfo
-	var unflushed []*datapb.SegmentInfo
-	var seekPosition *internalpb.MsgPosition
-	for _, s := range segments {
-		if (partitionID > allPartitionID && s.PartitionID != partitionID) ||
-			(s.GetStartPosition() == nil && s.GetDmlPosition() == nil) {
-			continue
-		}
-
-		if s.GetState() == commonpb.SegmentState_Flushing || s.GetState() == commonpb.SegmentState_Flushed {
-			flushed = append(flushed, trimSegmentInfo(s.SegmentInfo))
-		} else {
-			unflushed = append(unflushed, s.SegmentInfo)
-		}
-
-		var segmentPosition *internalpb.MsgPosition
-		if s.GetDmlPosition() != nil {
-			segmentPosition = s.GetDmlPosition()
-		} else {
-			segmentPosition = s.GetStartPosition()
-		}
-
-		if seekPosition == nil || segmentPosition.Timestamp < seekPosition.Timestamp {
-			seekPosition = segmentPosition
-		}
-	}
-	// use collection start position when segment position is not found
-	if seekPosition == nil {
-		collection := s.GetCollection(s.ctx, collectionID)
-		if collection != nil {
-			seekPosition = getCollectionStartPosition(channel, collection)
-		}
-	}
-
-	return &datapb.VchannelInfo{
-		CollectionID:      collectionID,
-		ChannelName:       channel,
-		SeekPosition:      seekPosition,
-		FlushedSegments:   flushed,
-		UnflushedSegments: unflushed,
-	}
-}
-
-func getCollectionStartPosition(channel string, collectionInfo *datapb.CollectionInfo) *internalpb.MsgPosition {
-	for _, sp := range collectionInfo.GetStartPositions() {
-		if sp.GetKey() != rootcoord.ToPhysicalChannel(channel) {
-			continue
-		}
-		return &internalpb.MsgPosition{
-			ChannelName: channel,
-			MsgID:       sp.GetData(),
-		}
-	}
-	return nil
-}
-
-// trimSegmentInfo returns a shallow copy of datapb.SegmentInfo and sets ALL binlog info to nil
-func trimSegmentInfo(info *datapb.SegmentInfo) *datapb.SegmentInfo {
-	return &datapb.SegmentInfo{
-		ID:             info.ID,
-		CollectionID:   info.CollectionID,
-		PartitionID:    info.PartitionID,
-		InsertChannel:  info.InsertChannel,
-		NumOfRows:      info.NumOfRows,
-		State:          info.State,
-		MaxRowNum:      info.MaxRowNum,
-		LastExpireTime: info.LastExpireTime,
-		StartPosition:  info.StartPosition,
-		DmlPosition:    info.DmlPosition,
-	}
-}
-
-func (s *Server) GetCollection(ctx context.Context, collectionID UniqueID) *datapb.CollectionInfo {
-	coll := s.meta.GetCollection(collectionID)
-	if coll != nil {
-		return coll
-	}
-	err := s.loadCollectionFromRootCoord(ctx, collectionID)
-	if err != nil {
-		log.Warn("failed to load collection from RootCoord", zap.Int64("collectionID", collectionID), zap.Error(err))
-	}
-
-	return s.meta.GetCollection(collectionID)
 }
